requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
Pillow==10.2.0
numpy==1.25.2
tensorflow==2.15.0
keras==2.15.0
plotly==5.24.1
gunicorn==21.2.0
h5py==3.10.0
